enable explicit branching map for wfxors
The `Data` item of the task definition can used for a branch map. A
map is created if not specified.

This also fixes the problem where the task orders in the `Succ` list
are lost when creating the digraph. `wfxors` is the only task type
where the order may matter.